make a util for opening a plot origin file
